feat(App): add modal functionality for adding goals
Add the ability to open and close a modal for adding goals. The modal is triggered by a button press and allows the user to enter a new goal. When the goal is added, the modal is closed. This enhances the user experience by providing a more focused and intuitive way to add goals.